Added WallScraper mini-game
This is a concept for the first minigame that has to be completed in
order to progress the puzzles; this one is scraping wallpaper.

There's currently no way to run this from the main game; launch the
WallScraper scene in Minigames/WallScraper to see it.  It also has no
success state, and the failure state is really rudamentary - but it
gets the idea across.